Remove exceptions module
